+++ core/client/app/routes/team/index.js
@@ -3,8 +3,7 @@ import CurrentUserSettings from 'ghost/mixins/current-user-settings';
 import PaginationRouteMixin from 'ghost/mixins/pagination-route';
 import styleBody from 'ghost/mixins/style-body';
 
-var paginationSettings,
-    TeamIndexRoute;
+var paginationSettings;
 
 paginationSettings = {
     page: 1,
@@ -12,10 +11,10 @@ paginationSettings = {
     status: 'active'
 };
 
-TeamIndexRoute = AuthenticatedRoute.extend(styleBody, CurrentUserSettings, PaginationRouteMixin, {
+export default AuthenticatedRoute.extend(styleBody, CurrentUserSettings, PaginationRouteMixin, {
     titleToken: 'Team',
 
-    classNames: ['view-users'],
+    classNames: ['view-team'],
 
     setupController: function (controller, model) {
         this._super(controller, model);
@@ -24,26 +23,14 @@ TeamIndexRoute = AuthenticatedRoute.extend(styleBody, CurrentUserSettings, Pagin
 
     beforeModel: function (transition) {
         this._super(transition);
-        return this.get('session.user')
-            .then(this.transitionAuthor());
     },
 
     model: function () {
         var self = this;
 
         return self.store.find('user', {limit: 'all', status: 'invited'}).then(function () {
-            return self.get('session.user').then(function (currentUser) {
-                if (currentUser.get('isEditor')) {
-                    // Editors only see authors in the list
-                    paginationSettings.role = 'Author';
-                }
-
-                return self.store.filter('user', paginationSettings, function (user) {
-                    if (currentUser.get('isEditor')) {
-                        return user.get('isAuthor') || user === currentUser;
-                    }
-                    return true;
-                });
+            return self.store.filter('user', paginationSettings, function () {
+                return true;
             });
         });
     },
@@ -54,5 +41,3 @@ TeamIndexRoute = AuthenticatedRoute.extend(styleBody, CurrentUserSettings, Pagin
         }
     }
 });
-
-export default TeamIndexRoute;
