+++ core/client/app/templates/team/index.hbs
@@ -1,73 +1,69 @@
 <section class="gh-view">
     <header class="view-header">
         {{#gh-view-title openMobileMenu="openMobileMenu"}}Team{{/gh-view-title}}
-        <section class="view-actions">
-            <button class="btn btn-green" {{action "openModal" "invite-new-user"}} >Invite People</button>
-        </section>
+        {{!-- Do not show Invite user button to authors --}}
+        {{#unless session.user.isAuthor}}
+            <section class="view-actions">
+                <button class="btn btn-green" {{action "openModal" "invite-new-user"}} >Invite People</button>
+            </section>
+        {{/unless}}
     </header>
 
     {{#gh-infinite-scroll
         fetch="loadNextPage"
         isLoading=isLoading
         tagName="section"
-        classNames="view-content settings-users"
+        classNames="view-content team"
     }}
-        {{#if invitedUsers}}
-            <section class="user-list invited-users">
-                <h4 class="user-list-title">Invited users</h4>
-                {{#each invitedUsers as |user|}}
-                    {{#gh-user-invited user=user reload="reload" as |component|}}
-                        <div class="user-list-item">
-                            <span class="user-list-item-icon icon-mail">ic</span>
-                            <div class="user-list-item-body">
-                                <span class="name">{{user.email}}</span><br>
-                                    {{#if user.model.pending}}
-                                        <span class="red">Invitation not sent - please try again</span>
-                                    {{else}}
-                                        <span class="description">
-                                            Invitation sent: {{component.createdAt}}
-                                        </span>
-                                    {{/if}}
+        {{!-- Do not show invited users to authors --}}
+        {{#unless session.user.isAuthor}}
+            {{#if invitedUsers}}
+                <section class="user-list invited-users">
+                    <h4 class="user-list-title">Invited users</h4>
+                    {{#each invitedUsers as |user|}}
+                        {{#gh-user-invited user=user reload="reload" as |component|}}
+                            <div class="user-list-item">
+                                <span class="user-list-item-icon icon-mail">ic</span>
+                                <div class="user-list-item-body">
+                                    <span class="name">{{user.email}}</span><br>
+                                        {{#if user.model.pending}}
+                                            <span class="red">Invitation not sent - please try again</span>
+                                        {{else}}
+                                            <span class="description">
+                                                Invitation sent: {{component.createdAt}}
+                                            </span>
+                                        {{/if}}
+                                </div>
+                                <aside class="user-list-item-aside">
+                                    <a class="user-list-action" href="#" {{action "revoke" target=component}}>
+                                        Revoke
+                                    </a>
+                                    <a class="user-list-action" href="#" {{action "resend" target=component}}>
+                                        Resend
+                                    </a>
+                                </aside>
                             </div>
-                            <aside class="user-list-item-aside">
-                                <a class="user-list-action" href="#" {{action "revoke" target=component}}>
-                                    Revoke
-                                </a>
-                                <a class="user-list-action" href="#" {{action "resend" target=component}}>
-                                    Resend
-                                </a>
-                            </aside>
-                        </div>
-                    {{/gh-user-invited}}
-                {{/each}}
-            </section>
-        {{/if}}
+                        {{/gh-user-invited}}
+                    {{/each}}
+                </section>
+            {{/if}}
+        {{/unless}}
 
         <section class="user-list active-users">
             <h4 class="user-list-title">Active users</h4>
             {{#each activeUsers key="id" as |user|}}
-                {{#gh-user-active user=user as |component|}}
-                    {{#link-to 'team.user' user class="user-list-item" }}
-                        <span class="user-list-item-figure" style={{component.userImageBackground}}>
-                            <span class="hidden">Photo of {{user.name}}</span>
-                        </span>
-
-                        <div class="user-list-item-body">
-                            <span class="name">
-                                {{user.name}}
-                            </span>
-                            <br>
-                            <span class="description">Last seen: {{component.lastLogin}}</span>
-                        </div>
-                        <aside class="user-list-item-aside">
-                            {{#unless user.isAuthor}}
-                                {{#each user.roles key="id" as |role|}}
-                                    <span class="role-label {{role.lowerCaseName}}">{{role.name}}</span>
-                                {{/each}}
-                            {{/unless}}
-                        </aside>
-                    {{/link-to}}
-                {{/gh-user-active}}
+                {{!-- For authors only shows users as a list, otherwise show users with links to user page --}}
+                {{#unless session.user.isAuthor}}
+                    {{#gh-user-active user=user as |component|}}
+                        {{#link-to 'team.user' user class="user-list-item"}}
+                            {{partial 'user-list-item'}}
+                        {{/link-to}}
+                    {{/gh-user-active}}
+                {{else}}
+                    {{#gh-user-active user=user as |component|}}
+                        <li class="ember-view active user-list-item">{{partial 'user-list-item'}}</li>
+                    {{/gh-user-active}}
+                {{/unless}}
             {{/each}}
         </section>
     {{/gh-infinite-scroll}}
